Passage Entité Integer en double

+++ src/Entity/FeeSheet.php
@@ -30,7 +30,7 @@ class FeeSheet
     private $nbDocuments;
 
     /**
-     * @ORM\Column(type="integer")
+     * @ORM\Column(type="float")
      */
     private $validAmount;
 
@@ -89,12 +89,12 @@ class FeeSheet
         return $this;
     }
 
-    public function getValidAmount(): ?int
+    public function getValidAmount(): ?float
     {
         return $this->validAmount;
     }
 
-    public function setValidAmount(int $validAmount): self
+    public function setValidAmount(float $validAmount): self
     {
         $this->validAmount = $validAmount;
