+++ src/Entity/VariableFeesLine.php
@@ -28,7 +28,7 @@ class VariableFeesLine
     private $date;
 
     /**
-     * @ORM\Column(type="integer")
+     * @ORM\Column(type="float")
      */
     private $amount;
 
@@ -66,12 +66,12 @@ class VariableFeesLine
         return $this;
     }
 
-    public function getAmount(): ?int
+    public function getAmount(): ?float
     {
         return $this->amount;
     }
 
-    public function setAmount(int $amount): self
+    public function setAmount(float $amount): self
     {
         $this->amount = $amount;
